refactor: move gas analytics helpers from gas-analytics-manager to helpers module
