+++ packages/hardhat/src/internal/builtin-plugins/gas-analytics/snapshot-cheatcodes.ts
@@ -18,8 +18,7 @@ import {
   parseFullyQualifiedName,
 } from "../../../utils/contract-names.js";
 
-import { getUserFqn } from "./gas-analytics-manager.js";
-import { formatSectionHeader } from "./helpers.js";
+import { formatSectionHeader, getUserFqn } from "./helpers.js";
 
 export const SNAPSHOT_CHEATCODES_DIR = "snapshots";
